Rename include/boost_helpers/buffer_pointer_converter.hpp -> buffer.hpp Rename src/buffer_pointer_converter.hpp -> buffer.cpp
Restructuring of buffer.hpp, moving the source code into the header.
buffer.cpp will be deleted on next commit.

Reduce the amount of code needed to wrap an iostream.
Now, all that is required is to instantiate a single template with a class
conforming to std::iostream.

 e.g.
boost::python::make_buffer_type_object<BufferObjectExample::Stream>();

That's about it. Not all buffer-like functions have been enabled on the
PyTypeObject yet. Going to split the template for istreams and ostreams
now. Finger's crossed that won't be tooo hard!